\name{eks-internal}
\alias{as_tidy_ks}
\alias{dim_ks}
\alias{unfactor}
\alias{label_percent}
\alias{tidy_ks}
\alias{scale_asymmetric}
\alias{scale_transp_neutral}
\alias{st_ks}
\alias{st_add_coordinates}
\alias{st_remove_long_segment}

\title{Internal functions in the eks library}

\description{These functions are user-level but which the user is not required to use directly.}

\value{The user is not required to use directly these outputs.}

\keyword{internal}
