\name{eks-internal}
\alias{as_tidy_ks}
\alias{dim_ks}
\alias{tidy_ks}
\alias{st_ks}
\alias{st_cast_ks}
\alias{st_remove_long_segment}
\title{Internal functions in the eks library}

\description{These functions are user-level but which the user is not required to use directly.}

\value{The user is not required to use directly these outputs.}

\keyword{internal}
